crosh: drop ssh support
Since we have a dedicated & secure extension, there's no need to let users
run the system copy of ssh.  This lets us drop a large body of code from
crosh that tried to validate the user input, as well as avoid running code
as the user that is not sandboxed.

BUG=chromium:468232
TEST=run unittests

Change-Id: Ie9b3a40b4a9370add34eded29e235cb5663eaf03
Reviewed-on: https://chromium-review.googlesource.com/260730
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Kees Cook <keescook@chromium.org>