installer: work around an sfdisk bug for hybrid MBR boards
When running chromeos-install on the SDK host (inside the chroot),
the sfdisk command sometimes fails inexplicably.  This only happens
for boards that have a "hybrid MBR" (because that's the only case
where sfdisk will be invoked).  This adds a check for the failure,
and works around it.

BUG=chromium:343681
TEST=Run the change on a system exhibiting the failure, see it pass

Change-Id: Ic3f95b8275242ff6999e693a7a38f047b04d8dd9
Reviewed-on: https://chromium-review.googlesource.com/260962
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>