power: send_metrics_on_resume: fix for veyron
None of the messages for the start of resume or the end of suspend show up on
veyron, so this adds one for that. It also simplifies the code for finding the
start time of resume and the end time of suspend since these are assumed to be
the same.

BUG=chrome-os-partner:35862
TEST=run power_Resume on jerry and sanity check results

Change-Id: Ibc2227e091293811513d2a35e8ecd73eb3a3958c
Reviewed-on: https://chromium-review.googlesource.com/261040
Reviewed-by: Dan Erat <derat@chromium.org>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>