shill: configure preferred device for wifi services
When a system has multiple wifi devices, it is desirable to allow users to
be able to configure wifi services to use specific wifi device for connecting.
A new wifi service configuration parameter "WiFi.PreferredDevice" is added for
this purpose.

With multiple wifi devices, a wifi endpoint will be created/added for each
BSS detected on each wifi device.  So when a wifi service tries to select an
endpoint to "use", it will select the endpoint with the best signal that's
associated with the preferred device if one exist, otherwise the endpoint with
the best signal overall.  This endpoint selection is mainly for determining
the device to use to connect to that service (meaning the device on which
"ConnectTo()" is performed if the service is connected), and the estimate of
the signal strength when connected.  wpa_supplicant is in charge for deciding
which endpoint to actually connect to.

There will be no side effect if the specified device is not an enabled
WiFi device.

BUG=chromium:467027
TEST=USE="asan clang" emerge-$BOARD shill
Manual Test:
1. Grab a ath9k device (wolf), and setup one additional wifi interface with:
   "iw phy phy0 interface add wlan1 type managed"
2. Change the MAC address of wlan1 to be different from wlan0:
   "ifconfig wlan1 hw ether [MAC]"
   then bring up the interface using "ifconfig wlan1 up"
3. Verify shill is managing wlan1 with:
   "/usr/local/lib/flimflam/test/list-devices
4. Configure the service with WiFi.PreferredDevice set using:
   "/usr/local/lib/flimflam/test/configure-service [ssid] ..."
5. Connect to the service using:
   "/usr/local/lib/flimflam/test/connect-service [ssid]"
6. Verify the device is connected to the service on the specified
   wifi interface.
CQ-DEPEND=CL:261049

Change-Id: I55838515b6f154d069ab5d02ded46e2f6148840c
Reviewed-on: https://chromium-review.googlesource.com/261210
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Zeping Qiu <zqiu@chromium.org>
Tested-by: Zeping Qiu <zqiu@chromium.org>