leaderd: Rename ChallengeLeader methods to LeaderChallenge
A leader challenge is a particular kind of message.  When we send this
message, we're not even always challenging a leader.  For instance, we
also use this message to discover who the leader is.

Add verbs where appropriate.  For instance, we now have
HandleLeaderChallenge and SendLeaderChallenge.

BUG=None
TEST=unittests still pass.

Change-Id: Ia3a6da26672907c80060fed9cb269d71553edbb2
Reviewed-on: https://chromium-review.googlesource.com/261192
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Commit-Queue: Christopher Wiley <wiley@chromium.org>
Tested-by: Christopher Wiley <wiley@chromium.org>