installer:  Fix EBUSY during BLKRRPART once and for all.
The previous fix wasn't effective on all code paths.  In some
cases, a later call to `blockdev --rereadpt` could fail with a
similar error.  The best diagnosis is that the call is in a race
with udev, so we add a call to `udevadm settle` to prevent the
conflict.

BUG=chromium:343681
TEST=run in chroot, run on a test device, try it in recovery

Change-Id: Ic2aa8f5377dbaa2b398cdc1c9e4d88b89fca54b9
Reviewed-on: https://chromium-review.googlesource.com/261356
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>