cryptohome: Send TPM dictionary attack counter metrics hourly.
This CL also changes the order in which the status of dictionary attack
reset status is reported. Now, if a counter is at 0 it will report
'reset not necessary' whether or not reset permissions exist. Previously
it would report 'reset not permitted' regardless of the counter value.

BUG=chromium:431360
TEST=unit, manually verified samples are firing

Change-Id: I0aa8f9450ccf36171ba0b08c96f050715acca5e2
Reviewed-on: https://chromium-review.googlesource.com/261626
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>
Commit-Queue: Darren Krahn <dkrahn@chromium.org>
Tested-by: Darren Krahn <dkrahn@chromium.org>