quipper: Add base class references to PerfParser and PerfSerializer
I would like to get rid of the inheritance between PerfSerializer =>
PerfParser => PerfReader, and replace it with composition. This change is
a transitional measure to assist that. Add a reference to the base classes
so that references to it can be gradually replaced with explicit
references. Later, the reference can be switched to a real object, and the
inheritance removed.

TEST=unit tests
BUG=none

Change-Id: I2248deb1e8251076afcd4e21fb6e353f28dddb92
Reviewed-on: https://chromium-review.googlesource.com/265914
Reviewed-by: Simon Que <sque@chromium.org>
Commit-Queue: David Sharp <dhsharp@chromium.org>
Tested-by: David Sharp <dhsharp@chromium.org>