permission_broker: Add CheckPathAccess method for filtering devices.
This method allows a client to check what the permission broker's
answer will likely be without actually granting permission for the
device. This will be used by Chrome to filter out devices that are
not accessible when prompting the user.

BUG=chromium:441526
TEST=Combined with https://codereview.chromium.org/1034333002 to test
that only allowed devices are visible to the user.

Change-Id: Ica029425bf089f1bd0b79adb20d9286ae2d5ea01
Reviewed-on: https://chromium-review.googlesource.com/262742
Tested-by: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: Reilly Grant <reillyg@chromium.org>