update_engine: Silence errors from CanResumeUpdate()
TEST_AND_RETURN_FALSE macro will log an ERROR when the value is false.
This macro is often used to bubble up errors. In the case of
CanResumeUpdate(), if the update can't be resumed (as happens with every
new update) it still logs a not so descriptive ERROR message.

This patch silence the LOG messages from CanResumeUpdate() calls.
There's already a log message for "Restarting" or "Starting" a new
update.

BUG=None
TEST=unittest still pass.

Change-Id: Iaf51c0a00f40ed9dd9cc96dcd5fcaabcb78049f4
Reviewed-on: https://chromium-review.googlesource.com/263133
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>