power: make suspend_stress_test work with nonexistent disable_dark_resume
Mounting a tempfile on top of /var/lib/power_manager/disable_dark_resume
fails if that pref does not already exist. This may leave dark resume
enabled during suspend_stress_test which will cause unintended behavior
as the RTC is likely to be a dark resume source.

This uses the /var/lib file directly if it doesn't already exist, and
bind-mounts a tempfile on top of it if it does. Both cases are cleaned
up appropriately.

BUG=chromium:453052,chrome-os-partner:38455
TEST=run suspend_stress_test with default prefs,
  run again with /usr/share/power_manager/board_specific/disable_dark_resume
    set to 0,
  run again with /var/lib/power_manager/disable_dark_resume set to 0,
  make sure there are no dark resumes along any pathway.

Change-Id: I5b152f29ae27f5875f77ebd9544e7a3368a26eda
Reviewed-on: https://chromium-review.googlesource.com/263137
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>
Commit-Queue: Eric Caruso <ejcaruso@chromium.org>