login: Add RestartJobWithAuth (again!)
The return of RestartJobWithAuth! This time, using a file descriptor
and SO_PEERCRED to get the pid of the calling process, since the
dbus-daemon is no longer able to correctly provide this data once
the session_manager and the browser are in a PID namespace together.

BUG=chromium:448821
TEST=new unit test; autotests in follow-on CL

Change-Id: Ib686f82eb9e3b2f0f8ed98e8068ff7e82527bec9
Reviewed-on: https://chromium-review.googlesource.com/267634
Trybot-Ready: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Will Drewry <wad@chromium.org>
Commit-Queue: Chris Masone <cmasone@chromium.org>