psyche: Add containers to psyched.
When psyched receives a RequestService request for a
not-yet-known service, make it ask somad for the
ContainerSpec providing that service and create objects to
track the container and all of the services it claims to
provide.

Actually starting containers via germd is not yet
implemented.

BUG=brillo:317
TEST=added a bunch of tests

Change-Id: I8dc249bf1824afdad091fc4feece6b2fa94847a5
Reviewed-on: https://chromium-review.googlesource.com/264150
Tested-by: Dan Erat <derat@chromium.org>
Reviewed-by: Chris Masone <cmasone@chromium.org>
Commit-Queue: Dan Erat <derat@chromium.org>
Trybot-Ready: Dan Erat <derat@chromium.org>