psyche: Remove ServiceInterface::State enum.
Remove an enum that tracked the state of a service, i.e.
"started" vs. "stopped". This was added when the
relationship between services and containers was less clear.

A Service object's lifetime now matches that of the
Container object that provides it, and the presence of a
binder proxy is sufficient to determine whether the service
is currently registered/callable or not.

BUG=none
TEST=unit tests pass and psyche_demo binaries still work

Change-Id: I2176ef31b7658a4ef5c716e9e47cc44c40658536
Reviewed-on: https://chromium-review.googlesource.com/264762
Tested-by: Dan Erat <derat@chromium.org>
Reviewed-by: Mike Colagrosso <mcolagrosso@chromium.org>
Commit-Queue: Dan Erat <derat@chromium.org>
Trybot-Ready: Dan Erat <derat@chromium.org>