bidl: change enum to strings for method selection
bidl used to define an enum for each rpc method
and switch() that method id to decide what method
to call. This causes problems with backwards compat
as these ids depend on the order the functions are
defined in the proto file, which may change.

Now the function name is used and a string match
decided what method is called.

BUG=brillo:732
TEST=pysche_demo

Change-Id: Ib6d4adbd564c638cebca742dc2d533dd8c4f5caf
Reviewed-on: https://chromium-review.googlesource.com/264957
Tested-by: Lee Campbell <leecam@chromium.org>
Trybot-Ready: Lee Campbell <leecam@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>
Commit-Queue: Lee Campbell <leecam@chromium.org>