soma: psyche: germ: Support multiple executables
We need to support ContainerSpecs that enumerate multiple
executables to be run inside a given container.
This CL rejiggers the ContainerSpec proto to add a submessage called
Executable that holds a commandline, uid/gid, and a few other things. It
removes those fields from the top-level ContainerSpec message.
ContainerSpecReader and ReadOnlyContainerSpec are updated to reflect these
changes, and the code in germd and psyched that references fields in
the ContainerSpec protobuf is updated as well

BUG=brillo:800
TEST=unit

Change-Id: I100b6a1ff36540975f85ecf7c0a28b817b051647
Reviewed-on: https://chromium-review.googlesource.com/265125
Reviewed-by: Chris Masone <cmasone@chromium.org>
Trybot-Ready: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Mike Colagrosso <mcolagrosso@chromium.org>
Commit-Queue: Chris Masone <cmasone@chromium.org>