cryptohome: Fix minor bug in LoadEnrollmentState
There is a problem with dbus-glib that if a function has a GArray** as a
parameter, and we don't set that value before returning, the DBus call
never returns an answer to the caller.

I added comments in all the other calls, to remind why that assignment has
to be at the beginning of the function.

BUG=None
TEST="Send dbus command:
      dbus-send --system \
                --dest=org.chromium.Cryptohome \
                --type=method_call --print-reply \
                /org/chromium/Cryptohome \
                org.chromium.CryptohomeInterface.LoadEnrollmentState
      without an enrollment_state.epb file"

Change-Id: I32ca87f6458ad44b0de70efb4b860bff9c5d821c
Reviewed-on: https://chromium-review.googlesource.com/266897
Reviewed-by: Darren Krahn <dkrahn@chromium.org>
Tested-by: Marco Vanotti <mvanotti@google.com>
Commit-Queue: Marco Vanotti <mvanotti@google.com>