init: Remove the hard-coded tty path when calling chromeos-boot-alert.
Some modes (warn_dev) in chromeos-boot-alert command need to read from TTY. This
may be a problem if given TTY file does not exist because we will also write
some text message to the path. In the end we will be reading the message we have
just outputted.

This is a serious problem for warn_dev because most messages contain space and
in warn_dev we consider space as "reboot in recovery mode", which makes
developer mode on devices using warn_dev (Alex, ZGB) totally unusable.

To prevent that, we should stop assigning /dev/tty1 as TTY path and let
chromeos-boot-alert to detect best console dynamically (especially that it may
need to call frecon in future to get a working pts), and fallback to /dev/null
if nothing is available (so at least it won't go to recovery mode).

BUG=chromium:479964
TEST=manual: chromeos-boot-alert warn_dev # won't trigger recovery mode.

Change-Id: I3c7dd4f0dc68ca2e8269d1f6cf085b3b5206f459
Reviewed-on: https://chromium-review.googlesource.com/266862
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>
Commit-Queue: Hung-Te Lin <hungte@chromium.org>