quipper: Correctly write out HEADER_BRANCH_STACK feature metadata
The change below caused a regression in which the size field of an entry
in the feature metadata index was written from uninintialized memory:

dd62cec quipper: (Const)BufferWithSize handles raw data and string reads/writes

HEADER_BRANCH_STACK does not have any metadata, but all features get an
entry in the feature metadata index. This oddity led to a misconception
that there is an extra perf_file_section at the end of the index that
points to the end of the file. Fix that misconception, and add a
regression test. Now all features have an index entry written.

TEST=added regression unit test
BUG=chromium:484393

Change-Id: I68f7bbb3736c6b6174cd7c0bb7a1895e9eb14495
Reviewed-on: https://chromium-review.googlesource.com/269563
Reviewed-by: David Sharp <dhsharp@chromium.org>
Commit-Queue: David Sharp <dhsharp@chromium.org>
Tested-by: David Sharp <dhsharp@chromium.org>