init: Fix TTY redirection when reading from console in chromeos-boot-alert.
When trying to read from TTY, we should first run stty to configure terminal
settings. Previously in chromeos-boot-alert:match_char_timeout we only read
from stdin and incorrectly call stty on stdout (which is not redirected), and
would cause a problem if system TTY default values are incorrect (for example,
if the STDOUT is a PTS created by frecon instead of VT /dev/tty1).

To fix this, we should pass TTY as an argument and specify that directly when
calling "stty" and "head".

Since we are actually using TTY for both input and output, the variable name
"STDOUT" is also changed to "TTY" to prevent confusion, and all redirection are
now using append (>>).

BUG=chromium:462460
TEST=chromeos-boot-alert power_wash

Change-Id: Icac74d24f5bcd2a7fe978552dafc9177405df8e5
Reviewed-on: https://chromium-review.googlesource.com/269612
Tested-by: Hung-Te Lin <hungte@chromium.org>
Commit-Queue: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: David Sodman <dsodman@chromium.org>