libchromeos: Add TLS stream support based on raw socket stream
chromeos::TlsStream is an SSL/TLS layer on top of raw socket stream.
TlsStream::Connect() performs TLS handshake on the socket and
establishes a secure connection to a server. Once this is done,
chromeos::TlsStream can be used to read from and write to the TLS
channel as if it was a normal socket. It will perform all the
necessary data encryption/decription and possible session
re-negotiations transparently in the background.

chromeos::TlsStream is fully asynchronous.

BUG=brillo:191
TEST=`FEATURES=test emerge-link libchromeos`

Change-Id: I62aa06e582a06926e236c85d743c30d31fad60e3
Reviewed-on: https://chromium-review.googlesource.com/269803
Reviewed-by: Alex Vakulenko <avakulenko@chromium.org>
Commit-Queue: Alex Vakulenko <avakulenko@chromium.org>
Tested-by: Alex Vakulenko <avakulenko@chromium.org>