lorgnette: Perform scan for Epson network-based scanners
The prober in sane-backends for Epson network-based scanners
does not listen to a well-known port.  The implementation is
simple enough to replicate in lorgnette, which can manipulate
permission_broker to gain access to a listening socket that can
receive the reply to broadcast probes.  Augment the information
from sane-backends with the result of this probe.

BUG=chromium:485827
TEST=Manual with Epson network-based scanner on the local LAN:
TEST=dbus-send --system --print-reply --fixed \
TEST=   --dest=org.chromium.lorgnette /org/chromium/lorgnette/Manager \
TEST=   org.chromium.lorgnette.Manager.ListScanners
TEST=Ensure "Epson" appears in the output

Change-Id: Id411c3cb117fdfad5fa6beef3ae03dbfd384351b
Reviewed-on: https://chromium-review.googlesource.com/270071
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>