crosh: rework history handling
Be proactive in sanitizing the history file in the user's home before
we load it directly into the shell.  This way if the file is corrupted
(either due to filesystem troubles of malicious code), we don't run the
risk of loading that into the shell.

We also set a few more shell variables to help limit the size of the
history file so it doesn't grow excessively large simply because it is
used a lot.

BUG=chromium:380538
TEST=loaded crosh with corrupt history files and made sure it was cleaned
TEST=loaded crosh with large history files and made sure it truncated

Change-Id: I54c5393c1cb3d3021b9d78818a509f1f00fdf50f
Reviewed-on: https://chromium-review.googlesource.com/271649
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Trybot-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>