libchromeos: fix libchromeos-test .pc file
Since libchromeos-test is a static archive, we need to depend directly on
the libs we use all the time and not just when --static is specified.

Similarly, we need to explicitly add libchromeos-test to the dep list.  We
can't do it via the dep list because the .pc file is generated at the same
time as the test lib.

BUG=chromium:493067
TEST=precq passes
TEST=building tests for mips works now

Change-Id: Ia3fa3a7bd5491e370c9dc6f1a051be788d7937fe
Reviewed-on: https://chromium-review.googlesource.com/273606
Reviewed-by: Alex Vakulenko <avakulenko@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>