trunks: skeleton for providing direct SPI access to TPM
This patch introduces an alternative TPM communications class
TrunksFtdiSpi, which can be used instead of the currently available
TrunksProxy.

Both these TPM communications classes are derived from
CommandTransceiver. The decision which one to use in made at the
TrunksFactoryImp class construction time, driven by the new Boolean
argument added to the TrunksFactoryImp constructor.

The Init() function had to be defined as a virtual member of the base
class, because TrunksFactoryImp expects the communicator to have this
function.

The SPI over FTDI interface is not available when compiling for the
boards. In order to avoid excessive #if/#else/#endif in the code, a
dummy implementation is provided for the board compilation case.

BUG=chromium:498861
TEST=emerging trunks succeeds for both host and board

Change-Id: Id398cc0c6a296b37e8301964c63f13cb1f4eae42
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/276823
Reviewed-by: Utkarsh Sanghi <usanghi@chromium.org>