shill: RTNLHandler: Mask known errors
Create a concept of an error mask -- a list of errno values that
are are considered benign for a given RTNL method call.  Maintain
window of these error masks for the last N most recent sequence
numbers sent to the kernel.  On receipt of an error response,
only send a LOG(ERROR) (visible in /var/log/messages) if this
error is not anticipated by the error mask.

Use this facility to create a variant of SendMessage() that
inspects the outgoing message request and assigns a default
error mask -- for example an "Add" request that receives an
EEXIST ("already exists") error is benign with respect to
method calls that just wish that this entry exist.

Use this new facility to make routing and address requests
silent in the face of requests that fail because the work
is already essentially done.

While here, make the normal processing of incoming and
outgoing RTNL messages at a less verbose level, since
libshill-net doesn't use context-specific logging.

BUG=chromium:498908
TEST=New unit tests (for regressions) + Manual:
TEST=restart shill; grep "shill.*rtnl" /var/log/messages

Change-Id: I9b4a9ae579dd34935a9c269979668dc569728790
Reviewed-on: https://chromium-review.googlesource.com/276739
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>