shill: DHCPCDProxy: Ignore messages from unbound PIDs
When an interface is destroyed, shill and dhcpcd race to destroy
configuration.  The situation that this CL resolves is one where
dhcpcd still has an outstanding signal that is delivered to the
DHCPCDProxy after the PID has been unbound from the DHCPProvider.

This CL maintains a short set of previously unbound PIDs that
can be used to dismiss messages from recently departed dhcpcd
instances without triggering an error message.

While here, correct the DHCPConfig unit test to use a mocked
DHCPProvider instead of a real one, and move testing of the
Bind and Unbind methods to the DHCPProvider unit test where it
makes more sense.

BUG=chromium:498908
TEST=Unit tests + "rmmod asix; modprobe asix" # Reload ethernet
TEST="grep 'Unknown DHCP client PID' /var/log/messages"

Change-Id: Ic55b0b2281651fd78a7e6fccb1b0fc2e7e520bcb
Reviewed-on: https://chromium-review.googlesource.com/277138
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>